requirements.txt
Flask
psycopg2-binary
Flask-Cors
requests
APScheduler
gunicorn